Fix checkbox 'isDuo' not updating
Problem: The form was correctly submitted and the new offer was passed as props to OfferCreation as expected.
However, the field 'isDuo' was not controlled. As a resut, the checkbox didn't refresh properly.
the property defaultChecked is used on initial render.
If you reloaded the page though, it would show the correct value.

Fix: use the property 'checked' which is controlled

Reference: https://reactjs.org/docs/forms.html#default-value